Move all internal methods to dedicated file instead of on QuickJs
