+++ zipline-loader/src/commonTest/kotlin/app/cash/zipline/loader/ZiplineLoaderTest.kt
@@ -17,6 +17,7 @@ package app.cash.zipline.loader
 
 import app.cash.turbine.test
 import app.cash.zipline.Zipline
+import app.cash.zipline.internal.getLog
 import app.cash.zipline.loader.internal.fetcher.LoadedManifest
 import app.cash.zipline.loader.internal.getApplicationManifestFileName
 import app.cash.zipline.loader.internal.systemEpochMsClock
@@ -75,7 +76,7 @@ class ZiplineLoaderTest {
     )
     val zipline = loader.loadOrFail("test", testFixtures.manifest)
     assertEquals(
-      zipline.quickJs.evaluate("globalThis.log", "assert.js"),
+      zipline.quickJs.getLog(),
       """
       |alpha loaded
       |bravo loaded
@@ -93,7 +94,7 @@ class ZiplineLoaderTest {
     )
     val zipline = (loader.loadOnce("test", manifestUrl) as LoadResult.Success).zipline
     assertEquals(
-      zipline.quickJs.evaluate("globalThis.log", "assert.js"),
+      zipline.quickJs.getLog(),
       """
       |alpha loaded
       |bravo loaded
@@ -154,7 +155,7 @@ class ZiplineLoaderTest {
     )
     val zipline = (loader.loadOnce("test", manifestUrl) as LoadResult.Success).zipline
     assertEquals(
-      zipline.quickJs.evaluate("globalThis.log", "assert.js"),
+      zipline.quickJs.getLog(),
       """
       |alpha loaded
       |bravo loaded
@@ -196,7 +197,7 @@ class ZiplineLoaderTest {
     // Load into Zipline.
     val zipline = loader.loadOrFail("test", testFixtures.manifest)
     assertEquals(
-      zipline.quickJs.evaluate("globalThis.log", "assert.js"),
+      zipline.quickJs.getLog(),
       """
       |alpha loaded
       |bravo loaded
