+++ zipline/src/jniMain/kotlin/app/cash/zipline/QuickJs.kt
@@ -15,17 +15,11 @@
  */
 package app.cash.zipline
 
-import app.cash.zipline.internal.COLLECT_DEPENDENCIES_DEFINE_JS
-import app.cash.zipline.internal.CURRENT_MODULE_DEPENDENCIES
-import app.cash.zipline.internal.CURRENT_MODULE_ID
-import app.cash.zipline.internal.DEFINE_JS
 import app.cash.zipline.internal.HostConsole
 import app.cash.zipline.internal.bridge.CallChannel
 import app.cash.zipline.internal.bridge.inboundChannelName
 import app.cash.zipline.internal.bridge.outboundChannelName
 import java.io.Closeable
-import kotlinx.serialization.decodeFromString
-import kotlinx.serialization.json.Json
 
 /**
  * An EMCAScript (Javascript) interpreter backed by the 'QuickJS' native engine.
@@ -128,47 +122,6 @@ actual class QuickJs private constructor(
     }
 
     return JniCallChannel(this, instance)
-  }
-
-  actual fun collectModuleDependencies() {
-    evaluate(COLLECT_DEPENDENCIES_DEFINE_JS, "collectDependencies.js")
-  }
-
-  actual fun getModuleDependencies(): List<String> {
-    val dependenciesString = getGlobalThis(CURRENT_MODULE_DEPENDENCIES)
-    val dependencies = Json.decodeFromString<List<String>>(
-      dependenciesString
-      // If define is never called, dependencies is returned as null
-        ?: "[]"
-    )
-    return dependencies
-  }
-
-  actual fun getGlobalThis(key: String): String? {
-    return evaluate("globalThis.$key", "getGlobalThis.js") as String?
-  }
-
-  internal actual fun initModuleLoader() {
-    evaluate(DEFINE_JS, "define.js")
-  }
-
-  internal actual fun loadJsModule(script: String, id: String) {
-    evaluate("globalThis.$CURRENT_MODULE_ID = '$id';")
-    evaluate(script, id)
-    evaluate("delete globalThis.$CURRENT_MODULE_ID;")
-  }
-
-  internal actual fun loadJsModule(id: String, bytecode: ByteArray) {
-    evaluate("globalThis.$CURRENT_MODULE_ID = '$id';")
-    execute(bytecode)
-    evaluate("delete globalThis.$CURRENT_MODULE_ID;")
-  }
-
-  actual fun runApplication(mainModuleId: String, mainFunction: String) {
-    evaluate(
-      script = "require('${mainModuleId}').$mainFunction()",
-      fileName = "RunApplication.kt",
-    )
   }
 
   actual fun gc() {
